Exercise 1 bug fixed 4

+++ tp3-arquitectura/Ejercicio 1/taylor.cpp
@@ -63,7 +63,7 @@ int main()
     long double local_result = calculate_term_process(x, rank * fraction_number, (rank + 1) * fraction_number);
 
     // Sum all results
-    MPI_Reduce(&local_result, &global_result, 1, MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+    MPI_Reduce(&local_result, &global_result, 1, MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
 
 
 
@@ -73,6 +73,7 @@ int main()
         cout << "Error finalizing MPI" << endl;
         exit(1);
     }
+
     cout << "El resultado es:" << 2.0 * global_result << endl;
 
     return 0;
